Changements BDD + gitignore
+ changement des montants de int à double

+++ src/main/java/fr/banque/entites/Transaction.java
@@ -20,6 +20,6 @@ public abstract class Transaction {
     @ManyToOne
     @JoinColumn(name = "compte_iban")
     private Compte compte;
-    private int montant;
+    private double montant;
     private String intitule;
 }
